refactor: move class out and fix imports
